+++ haskell/nettest/StablecoinClientTest.hs
@@ -12,15 +12,14 @@ import Morley.Nettest as NT
 import Tezos.Address (Address)
 import Util.Named ((.!))
 
-import Lorentz.Contracts.Stablecoin (stablecoinPermissionsDescriptor)
 import Stablecoin.Client
   (AddressAndAlias(..), Alias, InitialStorageData(..),
   UpdateOperatorData(AddOperator, RemoveOperator))
 import Stablecoin.Client.Cleveland
   (StablecoinScenario, acceptOwnership, assertEq, burn, changeMasterMinter, changePauser,
   configureMinter, deploy, getBalanceOf, getContractOwner, getMasterMinter, getMintingAllowance,
-  getPaused, getPauser, getPendingContractOwner, getTokenMetadata, getTotalSupply, getTransferlist,
-  isOperator, mint, pause, permissionsDescriptor, removeMinter, revealKeyUnlessRevealed,
+  getPaused, getPauser, getPendingContractOwner, getTokenMetadata, getTransferlist,
+  isOperator, mint, pause, removeMinter, revealKeyUnlessRevealed,
   setTransferlist, transferOwnership, unpause, updateOperators)
 import qualified Stablecoin.Client.Cleveland as SC
 
@@ -53,10 +52,6 @@ stablecoinClientScenario aliasPrefix = do
   expectedBalance <- NT.getBalance (AddressResolved contractAddr)
   actualBalance `assertEq` expectedBalance
 
-  comment "Testing permissions descriptor"
-  permissionsDescriptor >>= \pd ->
-    pd `assertEq` stablecoinPermissionsDescriptor
-
   comment "Testing set/get-transferlist"
   getTransferlist contract >>= \tl ->
     tl `assertEq` Just (AddressAndAlias transferlistAddr (Just transferlistAlias))
@@ -81,7 +76,6 @@ stablecoinClientScenario aliasPrefix = do
   burn (#sender .! minter) contract [1]
   getBalanceOf contract user1 >>=  \balance -> balance `assertEq` 7
   getBalanceOf contract user2 >>=  \balance -> balance `assertEq` 0
-  getTotalSupply contract >>= \supply -> supply `assertEq` 7
 
   SC.transfer (#sender .! minter) contract user1 user2 2
   getBalanceOf contract user1 >>=  \balance -> balance `assertEq` 5